Added a very very rough first pass of a physics "powered ragdoll" character. And he works with planets! Although there is a very off rotation bug when hes standing at the perfect 90degree (left side) of the planet, odd.
I added a prefab of him into a prefabs folder, just drag and drop it into a scene (or just mess with him on my test scene), it even includes a camera. Right now you have to assign the planet to the "test" variable on the character script, this is because i have to leave to eat dinner, ill make it auto find the nearest planet later today.